pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ask-ai-skills-builder"
version = "0.2.0"
description = "An interactive agent that discovers and interfaces with ASK AI features on developer documentation sites."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0"}
dependencies = [
    "playwright>=1.40.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "duckduckgo-search>=4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ask-ai-skills-builder"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
